--- src/mini_nn/models/arnav_mini_nn.py
@@ -1,21 +1,0 @@
-# arnav
-import torch.nn as nn
-
-
-class SimpleNet(nn.Module):
-    """
-    A simple feedforward neural network.
-    Architecture: Linear → ReLU → Linear
-    """
-    
-    def __init__(self, input_size, hidden_size, output_size):
-        super(SimpleNet, self).__init__()
-        self.fc1 = nn.Linear(input_size, hidden_size)
-        self.relu = nn.ReLU()
-        self.fc2 = nn.Linear(hidden_size, output_size)
-    
-    def forward(self, x):
-        x = self.fc1(x)
-        x = self.relu(x)
-        x = self.fc2(x)
-        return x
